requirements.txt
charset-normalizer==2.1.1
idna==3.4
mysql-connector==2.2.9
mysql-connector-python==8.0.31
numpy==1.24.1
pandas==1.5.2
protobuf==3.20.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
six==1.16.0
urllib3==1.26.13
wincertstore==0.2
xmltodict==0.13.0
